Fix: apply mentoring feedback
