+++ src/main/java/org/team6/coffeebeanery/order/controller/BuyerOrderController.java
@@ -2,7 +2,6 @@ package org.team6.coffeebeanery.order.controller;
 
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpHeaders;
@@ -11,12 +10,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.team6.coffeebeanery.common.dto.PageDTO;
-import org.team6.coffeebeanery.common.model.Address;
+import org.team6.coffeebeanery.order.dto.EmailReqBody;
+import org.team6.coffeebeanery.order.dto.OrderCreateReqBody;
 import org.team6.coffeebeanery.order.dto.OrderDTO;
-import org.team6.coffeebeanery.order.model.Order;
 import org.team6.coffeebeanery.order.service.BuyerOrderService;
 import org.team6.coffeebeanery.product.dto.ProductDTO;
-import org.team6.coffeebeanery.product.model.Product;
 import org.team6.coffeebeanery.product.service.BuyerProductService;
 
 import java.util.List;
@@ -27,37 +25,6 @@ import java.util.List;
 public class BuyerOrderController {
     private final BuyerOrderService orderService;
     private final BuyerProductService productService;
-
-    private static record EmailReqBody(
-            @NotBlank(message= "이메일은 필수입니다.")
-            @Email(message = "올바른 이메일 형식이 아닙니다")
-            String email) {
-    }
-
-    private static record OrderCreateReqBody(
-            @NotBlank(message = "이메일은 필수입니다")
-            @Email(message = "올바른 이메일 형식이 아닙니다")
-            String email,
-
-            @NotBlank(message = "기본 주소는 필수입니다")
-            @Size(max = 100, message = "기본 주소는 100자를 초과할 수 없습니다")
-            String baseAddress,
-
-            @NotBlank(message = "상세 주소는 필수입니다")
-            @Size(max = 100, message = "상세 주소는 100자를 초과할 수 없습니다")
-            String detailAddress,
-
-            @NotBlank(message = "우편번호는 필수입니다")
-            @Pattern(regexp = "\\d{5}", message = "우편번호는 5자리 숫자여야 합니다")
-            String zipCode,
-
-            @NotNull(message = "총 금액은 필수입니다")
-            Long totalPrice
-    ) {
-        public Address toAddress() {
-            return new Address(baseAddress, detailAddress, zipCode);
-        }
-    }
 
     // 이메일 검증
     @PostMapping("/email")
@@ -79,13 +46,14 @@ public class BuyerOrderController {
                                        @RequestParam(value = "page", defaultValue = "0") int page) {
 
         Page<OrderDTO> paging = orderService.getListByEmail(customerEmail, page);
-        return new PageDTO<OrderDTO>(paging);
+        return new PageDTO<>(paging);
     }
 
     // 장바구니 결제 수행
     @Transactional
     @PostMapping("")
-    public ResponseEntity<String> order(@Valid @RequestBody OrderCreateReqBody reqBody,
+    @ResponseStatus(HttpStatus.OK)
+    public void order(@Valid @RequestBody OrderCreateReqBody reqBody,
                                         HttpSession session) {
 
         List<ProductDTO> cart = (List<ProductDTO>)session.getAttribute("cart");
@@ -94,22 +62,10 @@ public class BuyerOrderController {
             throw new IllegalStateException("장바구니가 비어있습니다.");
         }
 
-        String email = reqBody.email();
-        Address address = reqBody.toAddress();
-
-        Order order = orderService.saveOrder(email, address, reqBody.totalPrice());
         try {
-            for (ProductDTO item : cart) {
-                Product product = productService.getProductById(item.getProductId());
-                orderService.saveOrderDetail(item, order, product);
-                productService.decreaseStock(item.getProductId(), item.getProductStock());
-            }
+            orderService.saveOrderWithCart(cart, reqBody);
         } catch(IllegalStateException e) {
             throw new IllegalStateException("주문 가능한 재고를 초과했습니다.");
         }
-
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body("주문 성공");
     }
 }
