+++ src/main/java/org/team6/coffeebeanery/order/controller/SellerOrderController.java
@@ -29,3 +29,4 @@ public class SellerOrderController {
         return new PageDTO<>(orderPage);
     }
 }
+
